ebayteam-exitcodes-8b3f811ef643/requirements_heroku.txt
-r requirements_base.txt
# packages that are unnecessary on Windows/Mac local dev environments, or don't install properly
psycopg2==2.5.1
django-toolbelt==0.0.1
pycrypto==2.6.1